backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
playwright==1.47.0
beautifulsoup4==4.12.3
firebase-admin==6.5.0
python-dotenv==1.0.1
pydantic==2.9.2
httpx==0.27.2
asyncio-throttle==1.0.2
